+++ Shared/LobbyClient/LobbyClient.csproj
@@ -78,6 +78,7 @@
     <Compile Include="ScriptGenerator.cs" />
     <Compile Include="Spring.cs" />
     <Compile Include="SpringieServer.cs" />
+    <Compile Include="SpringLogEventArgs.cs" />
     <Compile Include="SpringSettings.cs" />
     <Compile Include="Talker.cs" />
     <Compile Include="TasClient.cs" />
